Documentation adjustments and moving textformatter/tablemaker into ftext module
