+++ mfile
@@ -1,4 +1,3 @@
 {
-  "package": "MDK",
-  "version": "1"
+  "package": "MDK"
 }
