+++ src/resources/mdkversion.txt
@@ -1,1 +1,1 @@
-1.1.5
+2.0.0
