+++ src/scripts/MDKExample/Example.lua
@@ -1,10 +1,10 @@
-MDKExample = MDKExample or { version = "1.1.5" }
+MDKExample = MDKExample or { version = "2.0.0" }
 function MDKExample.exampleFText()
   local fText = require("@PKGNAME@.ftext")
   cecho(fText.fText("Testing!", {width = 40, formatType = 'c', textColor = '<orange>', capColor = '<purple>', cap = '[TEST]'}))
 end
 function MDKExample.exampleEMCO()
-  local EMCO = require("@PKGNAME@.EMCO")
+  local EMCO = require("@PKGNAME@.emco")
   MDKExample.UW = Geyser.UserWindow:new({name = "TestWindow"})
   local stylesheet = [[background-color: rgb(0,180,0,255); border-width: 1px; border-style: solid; border-color: gold; border-radius: 10px;]]
   local istylesheet = [[background-color: rgb(60,60,60,255); border-width: 1px; border-style: solid; border-color: gold; border-radius: 10px;]]
